+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Master
 ### Features
 - New option filter to attach lazily like `$(sel).on('click', filter, ...)`. Attaching is done for the whole set, not on each individual element.
 - New option otherClose to support extra closing buttons.
+- Gallery has new callbacks `beforeImage` and `afterImage`.
 
 ### Breaking changes
 - `$.featherlight` now always trigger a new lightbox. Use `$.fn.featherlight` for attaching events to elements.